+++ src/components/films-container.js
@@ -1,0 +1,19 @@
+
+/**
+ * Функция отрисовывает шаблон фильмов
+ * @return {string} Шаблон фильмов
+ */
+const createFilmsContainerTemplate = () => {
+  return (`
+  <section class="films">
+    <section class="films-list">
+      <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
+
+      <div class="films-list__container">
+
+      </div>
+    </section>
+  `.trim());
+};
+
+export {createFilmsContainerTemplate};
