+++ src/components/show-more-button.js
@@ -1,0 +1,9 @@
+/**
+ * Функция отрисовывает шаблон кнопки
+ * @return {string} Шаблон кнопки
+ */
+const createShowMoreBtnTemplate = () => {
+  return (`<button class="films-list__show-more">Show more</button>`.trim());
+};
+
+export {createShowMoreBtnTemplate};
