+++ src/components/top-rated.js
@@ -1,0 +1,17 @@
+/**
+ * Функция отрисовывает шаблон лучших фильмов
+ * @return {string} Шаблон лучших фильмов
+ */
+const createTopRatedTemplate = () => {
+  return (`
+  <section class="films-list--extra">
+      <h2 class="films-list__title">Top rated</h2>
+
+      <div class="films-list__container">
+
+      </div>
+    </section>
+  `.trim());
+};
+
+export {createTopRatedTemplate};
